lessons delete and study input database provider finished.

+++ lib/core/models/study_input_model/study_input.dart
@@ -7,7 +7,7 @@ class StudyInput extends DatabaseModel<StudyInput> {
   int questionNumber;
   int trueNumber;
   int falseNumber;
-  int netNumber;
+  double netNumber;
   int date;
 
   StudyInput(
